+++ Commands/WebParts/AddWebPartToWikiPage.cs
@@ -11,15 +11,16 @@ namespace OfficeDevPnP.PowerShell.Commands
     [CmdletHelp("Adds a webpart to a wiki page in a specified table row and column",
         Category = CmdletHelpCategory.WebParts)]
     [CmdletExample(
-Code = @"PS:> Add-SPOWebPartToWikiPage -PageUrl ""/sites/demo/sitepages/home.aspx"" -Path ""c:\myfiles\listview.webpart"" -Row 1 -Column 1",
+Code = @"PS:> Add-SPOWebPartToWikiPage -ServerRelativePageUrl ""/sites/demo/sitepages/home.aspx"" -Path ""c:\myfiles\listview.webpart"" -Row 1 -Column 1",
 Remarks = @"This will add the webpart as defined by the XML in the listview.webpart file to the specified page in the first row and the first column of the HTML table present on the page", SortOrder = 1)]
     [CmdletExample(
-  Code = @"PS:> Add-SPOWebPartToWikiPage -PageUrl ""/sites/demo/sitepages/home.aspx"" -XML $webpart -Row 1 -Column 1",
+  Code = @"PS:> Add-SPOWebPartToWikiPage -ServerRelativePageUrl ""/sites/demo/sitepages/home.aspx"" -XML $webpart -Row 1 -Column 1",
   Remarks = @"This will add the webpart as defined by the XML in the $webpart variable to the specified page in the first row and the first column of the HTML table present on the page", SortOrder = 2)]
     public class AddWebPartToWikiPage : SPOWebCmdlet
     {
         [Parameter(Mandatory = true, HelpMessage = "Full server relative url of the webpart page, e.g. /sites/demo/sitepages/home.aspx")]
-        public string PageUrl = string.Empty;
+        [Alias("PageUrl")]
+        public string ServerRelativePageUrl = string.Empty;
 
         [Parameter(Mandatory = true, ParameterSetName = "XML")]
         public string Xml = string.Empty;
@@ -62,7 +63,7 @@ Remarks = @"This will add the webpart as defined by the XML in the listview.webp
             }
             if (wp != null)
             {
-                SelectedWeb.AddWebPartToWikiPage(PageUrl, wp, Row, Column, AddSpace);
+                SelectedWeb.AddWebPartToWikiPage(ServerRelativePageUrl, wp, Row, Column, AddSpace);
             }
         }
     }