+++ Commands/WebParts/GetWebPart.cs
@@ -11,22 +11,23 @@ namespace OfficeDevPnP.PowerShell.Commands
     [CmdletHelp("Returns a webpart definition object",
         Category = CmdletHelpCategory.WebParts)]
     [CmdletExample(
-        Code = @"PS:> Get-SPOWebPart -PageUrl ""/sites/demo/sitepages/home.aspx""",
+        Code = @"PS:> Get-SPOWebPart -ServerRelativePageUrl ""/sites/demo/sitepages/home.aspx""",
         Remarks = @"Returns all webparts defined on the given page.", SortOrder = 1)]
     [CmdletExample(
-        Code = @"PS:> Get-SPOWebPart -PageUrl ""/sites/demo/sitepages/home.aspx"" -Identity a2875399-d6ff-43a0-96da-be6ae5875f82",
+        Code = @"PS:> Get-SPOWebPart -ServerRelativePageUrl ""/sites/demo/sitepages/home.aspx"" -Identity a2875399-d6ff-43a0-96da-be6ae5875f82",
         Remarks = @"Returns a specific webpart defined on the given page.", SortOrder = 2)]
     public class GetWebPart : SPOWebCmdlet
     {
         [Parameter(Mandatory = true, HelpMessage = "Full server relative url of the webpart page, e.g. /sites/mysite/sitepages/home.aspx")]
-        public string PageUrl = string.Empty;
+        [Alias("PageUrl")]
+        public string ServerRelativePageUrl = string.Empty;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ValueFromPipeline = true)]
         public WebPartPipeBind Identity;
 
         protected override void ExecuteCmdlet()
         {
-            var definitions = SelectedWeb.GetWebParts(PageUrl);
+            var definitions = SelectedWeb.GetWebParts(ServerRelativePageUrl);
 
             if (Identity != null)
             {