+++ Commands/WebParts/SetWebPartProperty.cs
@@ -11,7 +11,8 @@ namespace OfficeDevPnP.PowerShell.Commands
     public class SetWebPartProperty : SPOWebCmdlet
     {
         [Parameter(Mandatory = true)]
-        public string PageUrl = string.Empty;
+        [Alias("PageUrl")]
+        public string ServerRelativePageUrl = string.Empty;
 
         [Parameter(Mandatory = true)]
         public GuidPipeBind Identity;
@@ -26,11 +27,11 @@ namespace OfficeDevPnP.PowerShell.Commands
         {
             if (Value.BaseObject is string)
             {
-                SelectedWeb.SetWebPartProperty(Key, Value.ToString(), Identity.Id, PageUrl);
+                SelectedWeb.SetWebPartProperty(Key, Value.ToString(), Identity.Id, ServerRelativePageUrl);
             }
             else if (Value.BaseObject is int)
             {
-                SelectedWeb.SetWebPartProperty(Key, (int)Value.BaseObject, Identity.Id, PageUrl);
+                SelectedWeb.SetWebPartProperty(Key, (int)Value.BaseObject, Identity.Id, ServerRelativePageUrl);
             }
         }
     }