+++ changelog.md
@@ -1,5 +1,10 @@
 # OfficeDevPnP.PowerShell Changelog #
 
+**2015-09-21**
+
+* Added -Parameters parameter to Apply-SPOProvisioningTemplate. See help for the cmdlet for more info.
+* Renamed PageUrl parameter of Webpart cmdlets to ServerRelativePageUrl. Set PageUrl as parameter alias to not break existing scripts.
+
 **2015-09-17**
 
 * Added Get-SPOProperty to dynamically load specified properties from objects.